+++ CHANGELOG.md
@@ -1,2 +1,2 @@
 ## 0.1.0 (unreleased)
-- (...)
+- Add `Pdfkiwi.convertHtml()` method.